Fixed event log records leaked data between different classes
`cattr_accessor` register the accessor on a parent class, not on this
class. This meant that both GroupRegistered and ProductAddedToInventory
shared the list of data attributes. This didn't make sense because
GroupRegistered doesn't need a description.

Instead, key the attributes by the event log record class to only use
the attributes that are needed.